requirements.txt
flask
requests
python-dotenv
plotly
dash
dash-bootstrap-components
pandas
numpy
websocket-client
gunicorn



